Decompose builder orchestration files into focused modules
